Give the `Renderer` trait more of a reason to be
